custom editor complete

+++ Shooty-Blocks/Assets/Editor/EditorList.cs
@@ -2,19 +2,6 @@ using UnityEditor;
 using UnityEngine;
 using System;
 
-[Flags]
-public enum EditorListOption
-{
-    None = 0,
-    ListSize = 1,
-    ListLabel = 2,
-    ElementLabels = 4,
-    Buttons = 8,
-    Default = ListSize | ListLabel | ElementLabels,
-    NoElementLabels = ListSize | ListLabel,
-    All = Default | Buttons
-}
-
 public static class EditorList
 {
     private static GUIContent
@@ -26,25 +13,24 @@ public static class EditorList
 
     private static GUILayoutOption miniButtonWidth = GUILayout.Width(20f);
 
-    public static void Show(SerializedProperty list, SerializedProperty test)
+    public static void Show(SerializedProperty list, SerializedProperty difficultyBalance)
     {
-        EditorGUILayout.PropertyField(list);
-        for (int i = 0; i < list.arraySize; i++)
+        for (int i = list.arraySize-1; i >= 0; i--)
         {
-            //SerializedProperty difficultyBalance = blockGroup.FindPropertyRelative("m_difficultyBalance");
-
-            ShowElement(list, i, test);
+            ShowElement(list, i, difficultyBalance);
         }
 
         EditorGUILayout.BeginHorizontal();
         if (GUILayout.Button(addButtonContent, EditorStyles.miniButton))
         {
             list.arraySize += 1;
+            difficultyBalance.arraySize += 1;
         }
 
-        if (GUILayout.Button(removeButtonContent, EditorStyles.miniButton))
+        if (GUILayout.Button(removeButtonContent, EditorStyles.miniButton) && list.arraySize > 0)
         {
             int oldSize = list.arraySize;
+            difficultyBalance.DeleteArrayElementAtIndex(list.arraySize - 1);
             list.DeleteArrayElementAtIndex(list.arraySize-1);
             if (list.arraySize == oldSize)
             {
@@ -54,38 +40,47 @@ public static class EditorList
         EditorGUILayout.EndHorizontal();
     }
 
-    private static void ShowElement(SerializedProperty list, int i, SerializedProperty test)
+    private static void ShowElement(SerializedProperty list, int i, SerializedProperty difficultyBalance)
     {
+        
         SerializedProperty blockGroup = list.GetArrayElementAtIndex(i);
         
-
         EditorGUILayout.BeginHorizontal();
         EditorGUILayout.PropertyField(blockGroup, GUIContent.none);
-        ShowButtons(list, i);
+        int oldLength = list.arraySize;
+        ShowButtons(list, i, difficultyBalance);
         EditorGUILayout.EndHorizontal();
 
-        SerializedObject newObj = new SerializedObject(blockGroup.objectReferenceValue);
+        if (list.arraySize < oldLength)
+            return;
 
-        EditorGUILayout.PropertyField(newObj.FindProperty("m_layout"));
-        EditorGUILayout.PropertyField(newObj.FindProperty("m_difficultyBalance"));
+        if (list.GetArrayElementAtIndex(i).objectReferenceValue != null || blockGroup.objectReferenceValue != null)
+        {
+            SerializedObject newObj = new SerializedObject(blockGroup.objectReferenceValue);
 
+            EditorGUILayout.PropertyField(newObj.FindProperty("m_layout"));
+            EditorGUILayout.PropertyField(difficultyBalance.GetArrayElementAtIndex(i), new GUIContent("Difficulty Balance"));
+        }
         EditorGUILayout.Space(10);
     }
 
-    private static void ShowButtons(SerializedProperty list, int index)
+    private static void ShowButtons(SerializedProperty list, int index, SerializedProperty difficultyBalance)
     {
         if (GUILayout.Button(moveButtonContent, EditorStyles.miniButtonLeft, miniButtonWidth))
         {
             list.MoveArrayElement(index, index + 1);
+            difficultyBalance.MoveArrayElement(index, index + 1);
         }
         if (GUILayout.Button(duplicateButtonContent, EditorStyles.miniButtonMid, miniButtonWidth))
         {
             list.MoveArrayElement(index, index - 1);
+            difficultyBalance.MoveArrayElement(index, index - 1);
         }
         if (GUILayout.Button(deleteButtonContent, EditorStyles.miniButtonRight, miniButtonWidth))
         {
             int oldSize = list.arraySize;
             list.DeleteArrayElementAtIndex(index);
+            difficultyBalance.DeleteArrayElementAtIndex(index);
             if (list.arraySize == oldSize)
             {
                 list.DeleteArrayElementAtIndex(index);
